Re-enable the countdown script
Now everything is using UNIX epoch timestamps all over the plae it's a
lot easier to reason about what might / might not be going wrong -- and
in fact over the course of refactoring everything to use UNIX
timestamps, the original issues went away.

Although at time of writing we in the UK are in the +0000 timezone,
one of the biggest advantages of which is masking stupid issues in time
calculations -- at least for half of the year.