Extract more submissions code out, add a test
Using an SQLite in-memory database to test the user filtering, which
is probably going to need rewriting soon to support user roles, so
having a proper suite of tests for this filtering (eventually) should
make that less painful.

At the moment the database mock is created using a special argument
to db(inject_mock => xxx). There may be a nicer way of doing this the
Catalyst way. I should find out.